alter quantity field in InventoryItem model to FloatField and create migration

+++ interactive-kitchen-backend/inventory/models.py
@@ -2,7 +2,7 @@ from django.db import models
 
 class InventoryItem(models.Model):
     name = models.CharField(max_length=255)
-    quantity = models.PositiveIntegerField()
+    quantity = models.FloatField()
     unit = models.CharField(max_length=50)
     expiration_date = models.DateField(null=True, blank=True)
     added_by = models.ForeignKey('users.CustomUser', on_delete=models.CASCADE, related_name='inventory_items')
